third_party/blink/web_tests/external/wpt/pointerevents/pointerevent_after_target_removed.html

<!DOCTYPE HTML>
<title>Enter/leave events fired to parent after child is removed</title>
<meta name="variant" content="?mouse">
<meta name="variant" content="?touch">
<meta name="variant" content="?pen">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-actions.js"></script>
<script src="/resources/testdriver-vendor.js"></script>
<script src="pointerevent_support.js"></script>

<style>
  div.target {
      width: 100px;
      height: 100px;
  }
</style>
<div class="target" id="parent">
  <div class="target" id="child">child</div>
</div>
<div id="done">done</div>

<script>
  'use strict';
  const pointer_type = location.search.substring(1);

  const parent = document.getElementById("parent");
  const child = document.getElementById("child");
  const done = document.getElementById("done");

  let event_log = [];

  function logEvent(e) {
    if (e.eventPhase == e.AT_TARGET) {
      event_log.push(e.type + "@" + e.target.id);
    }
  }

  function removeChild() {
    parent.removeChild(child);
    event_log.push("(child-removed)");
  }

  function setup() {
    const logged_events = [
      "pointerover", "pointerout", "pointerenter", "pointerleave",
      "pointerdown", "pointerup"
    ];
    let targets = document.getElementsByClassName("target");
    for (let i = 0; i < targets.length; i++) {
      logged_events.forEach(ename => {
        targets[i].addEventListener(ename, logEvent);
      });
    }
  }

  function addPromiseTest(remover_event, expected_events) {
    const test_name = `PointerEvents from ${pointer_type} `+
        `received before/after child removal at ${remover_event}`;

    promise_test(async test => {
      event_log = [];

      child.addEventListener(remover_event, removeChild);
      test.add_cleanup(() => {
        child.removeEventListener(remover_event, removeChild);
        if (!child.parentElement) {
          parent.appendChild(child);
        }
      });

      let done_click_promise = getEvent("click", done);

      let actions = new test_driver.Actions()
          .addPointer("TestPointer", pointer_type)
          .pointerMove(-30, -30, {origin: parent})
          .pointerDown()
          .pointerUp()
          .pointerMove(30, 30, {origin: parent})
          .pointerDown()
          .pointerUp()
          .pointerMove(0, 0, {origin: done})
          .pointerDown()
          .pointerUp();

      await actions.send();
      await done_click_promise;

      assert_equals(event_log.toString(), expected_events.toString(),
          "events received");
    }, test_name);
  }

  setup();
  addPromiseTest("pointerdown", [
    "pointerover@child", "pointerenter@parent", "pointerenter@child",
    "pointerdown@child", "(child-removed)", "pointerup@parent",
    "pointerdown@parent", "pointerup@parent",
    "pointerout@parent", "pointerleave@parent"
  ]);
  addPromiseTest("pointerup", [
    "pointerover@child", "pointerenter@parent", "pointerenter@child",
    "pointerdown@child", "pointerup@child", "(child-removed)",
    "pointerdown@parent", "pointerup@parent",
    "pointerout@parent", "pointerleave@parent"
  ]);
</script>
